add user login check and update account fragment
The changes introduce a user login check in `HomeActivity` before loading the `AccountFragment`. If the user is not logged in, it redirects to the `LoginActivity`.
`AccountFragment` is updated to display user information, including name, gender, and date of birth.
Also add the navigation for account information, manage post, and change password.